General/requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.6
certifi==2018.8.24
cffi==1.14.6
chardet==3.0.4
cryptography==2.6.1
cson==0.7
distro==1.3.0
entrypoints==0.3
greenlet==1.1.1
httplib2==0.11.3
idna==2.6
Jinja2==2.10
keyring==17.1.1
keyrings.alt==3.1.1
MarkupSafe==1.1.0
mysql-connector-python==8.0.26
numpy==1.16.2
protobuf==3.17.3
pycparser==2.20
pycrypto==2.6.1
pymongo==3.12.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
pyxdg==0.25
requests==2.21.0
SecretStorage==2.3.1
six==1.12.0
speg==0.3
SQLAlchemy==1.4.22
urllib3==1.24.1
